add direct gauss motion

+++ slae/src/main/kotlin/method/DirectMethod.kt
@@ -6,7 +6,7 @@ import Matrix
  * @author Natalia Nikonova
  */
 interface DirectMethod {
-    fun getDeterminant() : Int
+    fun getDeterminant() : Double
     fun getTriangularMatrix() : Matrix
     fun getSolution() : List<Double>
     fun getInconsistencies() : List<Double>
